requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
scikit-learn==1.3.1
flask==3.0.0
pandas==2.1.1
WTForm==1.0
Flask-WTF==1.2.1
schedule==1.2.1